Classification of ASD (Images - Kaggle Dataset)
Autism is an issue that isn't generally analyzed in youngsters until around the age of 4 years. However, half of the parents of a child with autism report that they presumed an issue before their kid was 1 year old enough, and most guardians at first express worry to their pediatrician when their kid is year and a half old enough. This postpone implies that exceptionally small kids with autism are seldom perceived by experts. In this way, the early advancement of youngsters with autism stays a very remarkable secret. ASD is a huge neurodevelopmental problem that hinders a kid's capacity to convey and communicate with others. Ordinarily, perceiving a kid with ASD needs the determination by proficient specialists. In any case, it isn't just costly and tedious, yet additionally the outcomes are impacted by emotional components, like the experience of a specialist. This issue can be wiped out with the assistance of present day innovation and AI calculations. The utilization of machine learning calculations can incredibly affect the medical care area and result in early discovery of such issues such that ideal medical care can be given. In this project, the goal is to introduce different machine learning methods to distinguish if a kid is experiencing ASD. As the number of ASD cases are increasing it has therefore become an important task to detect it at an early stage so that there is a better chance of recovery. Techniques of Transferred Learning and Image Processing are a few of the tools used in order to tackle the fore mentioned problem.